Manejo de excepcion anadido para borrar persona cuando no encuentra ID

+++ src/main/java/com/sofkau/crudPersona/servicios/ServiciosPersona.java
@@ -35,6 +35,9 @@ public class ServiciosPersona implements InterfazServicioPersona{
 
     @Override
     public void borrar(int id) {
+        if (data.findById(id).isEmpty()){
+            throw new PersonaNoEncontrada();
+        }
         data.deleteById(id);
     }
